adding modules for news feed
