+++ index.html
@@ -4,22 +4,38 @@
 <head>
   <meta charset="UTF-8" />
   <link rel="icon" type="image/svg+xml" href="/favicon.svg" />
+  <link rel="stylesheet" href="./src/style.css">
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
   <title>Vite App</title>
 </head>
 
 <body>
   <div id="app">
-    <div class="news">
-      <select class="news-options">
-        <option value="+food">Food</option>
-        <option value="+wine">Wine</option>
-        <option value="+restuarants">Restuarants</option>
-      </select>
-    </div>
+    <header>
+      <nav>
+        <ul>
+          <li>Home</li>
+          <li>Meal Planning</li>
+          <li>Shopping List</li>
+          <li>Wines</li>
+          <li>Fast Food</li>
+          <li>Random Recipe</li>
+        </ul>
+      </nav>
+    </header>
+    <main>
+      <div class="news">
+        <!-- <select class="news-options">
+          <option value="+food">Food</option>
+          <option value="+wine">Wine</option>
+          <option value="+restuarants">Restuarants</option>
+        </select> -->
+      </div>
+    </main>
+
   </div>
   <script type="module" src="./src/main.js"></script>
-  <script src="./news_app.js"></script>
+  <script type="module" src="./src/news_app/news_app.js"></script>
 </body>
 
 </html>
